# perf(utils/url): Short circuit the regex in checkOptionalParameter