--- a/src/utils/url.ts
+++ b/src/utils/url.ts
@@ -166,7 +166,7 @@ export const checkOptionalParameter = (path: string): string[] | null => {
    in other cases it will return null
   */
 
-  if (!path.match(/\:.+\?$/)) {
+  if (path.charCodeAt(path.length - 1) != 63 || !path.match(/\:.+\?$/)) {
     return null
   }