@yusukebe 2025-03-04
Hi @csainty 

Awesome! But I would like to confirm. Is the benchmark script that you used `benchmarks/routers/src/bench-includes-init.mts`, not `bench.mts`?