@usualoma 2025-03-04
Hi, @csainty. Thank you. Good improvement!

I want to use `!==` as the comparison operator, and if I'm going to use this pattern, I want to replace subsequent condition to be `.includes(':')`